Extract incremental rebuild decision logic to KotlincModeFactory
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This diff
This diff relocates the logic determining whether to force non-incremental mode in the Incremental compiler from `JvmCompilationConfigurationFactory.kt` to `KotlincModeFactory`. By doing so, we reduce the amount of data passed down the flow and simplify the implementation of configuration detection, which will be introduced in subsequent changes.
 ---
**Internal**
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

Differential Revision: D72241030

fbshipit-source-id: 69743e20b57b99fae4a82cbfc17847dd84461419