Make java.base.jmod (or relevant proguard lib jar) accessible via java_toolchain
Summary:
Making the relevant libraryjar available from the jdk artifact so proguard can fetch it

# But jmods have native files!
Yes, but proguard only needs the class files, not the native artifacts, and there's no clean way to provide different artifacts. And from the advice of an expert in the space (IanChilds),

> Basically, you don't want a select to change some behavior of your toolchain depending upon which platform you're running on, because then your build won't be identical across platforms. Maybe not a huge deal, but we try to avoid that for caching reasons

So we're going with storing the asset with a name associated to its purpose and a warning comment not to use it for native purposes because it might not do what you want. These assets came from https://adoptium.net/temurin/releases/?variant=openjdk17&jvmVariant=hotspot&version=8 just like the jdk artifacts used in the dotslash references nextdoor.

# Well then if they're there in dotslash, why not use those?

Dotslash isn't used for that. It's kind of an OS-agnostic executable manager, but while it manages executable artifacts by OS, it doesn't manage arbitrary artifacts, [as they call out in their documentation](https://dotslash-cli.com/docs/limitations/#dotslash-is-not-a-package-manager)

# Fine but how about using CAS?
I mean sure, I could do that, but short of building my own fork of dotslash or at least its schema parser I can't do platform-independent builds. But if you really want CAS uploads I can figure that out. Push me on it, and I'll come back to it in two weeks to follow up and fix it. Pinky swear.

allow-large-files

Reviewed By: IanChilds

Differential Revision: D72294098

fbshipit-source-id: ea8e4d50ee149f613eb1e96fc8897a842a595079