xcframeworks | provider for non-swift
Summary:
# Context
Currently, when building an Apple .framework or .xcframework that contains Objective-C, you need to manually create a module.modulemap file and include it.

This is a minor pain in most circumstances, but is a bigger issue when we want to enable the easy production of frameworks for arbitrary targets (e.g. in the '[showcaseable](https://www.internalfb.com/code/fbsource/[e15875b807617ff1f566d108ee6e11568032da4b]/tools/build_defs/apple/showcase/showcaseable.bzl?lines=10-14)' system for IG )

There is already code to generate modulemaps for use in general compilation, this stack extends it to generate working modulemaps for embedding in frameworks.

# This Diff

AppleLibraryForDistributionInfo is currently used to represent the info for (Swift) libraries being packaged for distribution (i.e. in a framework).

This diff makes some of the fields optional so that it can also represent a modular Obj-C target for distribution (we need to use the module_name field).

Differential Revision: D72066625

fbshipit-source-id: 66936b39fdb971ef7eb01b72a970ed51d520117f