xcframeworks | add module_map = "auto" option
Summary:
# Context
Currently, when building an Apple .framework or .xcframework that contains Objective-C, you need to manually create a module.modulemap file and include it.

This is a minor pain in most circumstances, but is a bigger issue when we want to enable the easy production of frameworks for arbitrary targets (e.g. in the '[showcaseable](https://www.internalfb.com/code/fbsource/[e15875b807617ff1f566d108ee6e11568032da4b]/tools/build_defs/apple/showcase/showcaseable.bzl?lines=10-14)' system for IG )

There is already code to generate modulemaps for use in general compilation, this stack extends it to generate working modulemaps for embedding in frameworks.

# This Diff

Adds "auto" as an option to the module_map field in apple_bundle. It won't do anything yet

Differential Revision: D72066631

fbshipit-source-id: a5587827cf1b0ed8ade04940c762b4761534652e