xcframeworks | framework support in make_modulemap.py
Summary:
# Context
Currently, when building an Apple .framework or .xcframework that contains Objective-C, you need to manually create a module.modulemap file and include it.

This is a minor pain in most circumstances, but is a bigger issue when we want to enable the easy production of frameworks for arbitrary targets (e.g. in the '[showcaseable](https://www.internalfb.com/code/fbsource/[e15875b807617ff1f566d108ee6e11568032da4b]/tools/build_defs/apple/showcase/showcaseable.bzl?lines=10-14)' system for IG )

There is already code to generate modulemaps for use in general compilation, this stack extends it to generate working modulemaps for embedding in frameworks.

# This Diff

Supports building a modulemap for a framework in make_modulemap.py

We pass in a flag (`--framework`) and this changes the generated modulemap from one intended for use during a buck build:
1) It adds the string 'framework' to generate output like: `framework module Foo {}`
2) We don't prepend the headers with the internal buck paths

Differential Revision: D72066628

fbshipit-source-id: 393a5166d5ed9dbb510a78ddfe319f6f37415600