Add access_keyguard as role-level permission
Summary:
- this allows apps to use roles to grant this permission

# Context

For an app to register/restore a key, it needs access to the user pin. This access is gated by the permission com.meta.wrist.os.keyguardmanager.ACCESS_KEYGUARD (https://fburl.com/code/izzc4gc5). The issue is that this permission is signature|privileged, meaning that Android will only grant it to privapp apps.

Selfcare is a privapp, workout is not. I am not sure about other health apps and their privilege levels. This means we should be able to do full e2ee in selfcare, but not in workout

Differential Revision: D72407071

fbshipit-source-id: aaae19940f3e5433a6e0393a672b6c9c5c7504fa