introduce module_prefix. (#60)
If BUCK file is located inside a package source tree, then the earlier part of module source paths are truncated, and thus the resultant pkg.conf will have wrong information.
Though it is desirable to get correct module names, they are only correctly available as a result of parsing, i.e. needs to be implemented as dynamic values, which can complicate the situation. So for now (or forever), I assume that user can provide module prefix for that target without difficulty. Interface path is then adjusted with the prefix.