fix corner case in test shell when test node crashes during runs
Summary:
This change fixes type annotations in ct_daemon to reflect node_down events, and fixes a corner case in the test shell.

Currently, we crash the test run, when the test node disappears during a run. This has an unfortunate interaction with the shells debug mode an dit produces a crash dump.

This happens consistently, when we use a debugger, and end the debug session during a test run.

Reviewed By: jcpetruzza

Differential Revision: D72709762

fbshipit-source-id: 7177c9ce1d81ec460f787901cb8a1308f35c983f