Add plugin param to take class output folder. 1/n
Summary:
#facebook

## Context
Part of Kosabi 2.0 plan is changing stubgen to emit class bytecode directly for better performance.

rootfoo

 [[Original Proposal](https://fb.workplace.com/groups/2840058936242210/permalink/3743214705926624/)]

This aim to improve the performance of Kosabi stubgen step, and also with not output KT files, we avoid having to analzye these stub types in KSP and jvm-abi-gen step.

## This diff
Add a stubgen param to take a path param for where to emit the stub class

Reviewed By: nghuiqin

Differential Revision: D71973434

fbshipit-source-id: f6e1295647e5aef834ef84739e9eb4b0689d2267