Allow rustc_flags to take Artifact as argument
Summary:
When using Buck2 to bootstrap the Rust compiler, the ability to create a sysroot directory is needed (i.e., to collect rustc binary and standard libraries structurally into a single directory). We refer to such a target as *a sysroot target*, and the resulting directory as *a sysroot artifact*.

When creating a Rust toolchain from a sysroot artifact, compiler flags `["--sysroot", sysroot_artifact]` should be set. The current implementation, however, doesn't allow an `Artifact` as rustc_flags argument. This PR extends it to allow an `Artifact` for rustc_flags.

Here is an example of creating a sysroot (steps 1 through 3) and creating a Rust toolchain based on sysroot artifacts (steps 4 through 6), along with the error message I encountered (step 7):

{F1976949628}

X-link: https://github.com/facebook/buck2/pull/896

Reviewed By: JakobDegen

Differential Revision: D72749420

Pulled By: dtolnay

fbshipit-source-id: 3e553ddf86ad6d28c2f04ea9e29294fa26a0dc60