Consider IncrementalMetadata when determining if rebuild is necessary
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This stack
This stack aims to implement a mechanism that forces KIC to run non-incrementally in specific scenarios, such as when breaking changes are made to certain metadata files.

### This diff
This diff implements the logic of forcing a full recompilation when there are changes in incremental metadata.
----
**Internal**
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

Differential Revision: D72729635

fbshipit-source-id: ba939f32dafc8f974bf9ec8aeecd16776cd1e48f