Propagate incremental metadata digest into ActionMetadata component
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This stack
This stack aims to implement a mechanism that forces KIC to run non-incrementally in specific scenarios, such as when breaking changes are made to certain metadata files.

### This diff
This diff introduces filtering capabilities to ActionMetadata and incorporates incremental metadata digest as an attribute.

----
**Internal**
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

Differential Revision: D72726647

fbshipit-source-id: 73258f3eed760bad4494eee6031e72dae7649e79