Add core_build_mode to cfg name
Summary:
**NOTE**: Please ask if you need to revert this diff. This diff does not change build semantically. It updates the configuration name, which is an input to action key, meaning most previously cached build actions need to be rerun once becuase they have new action keys. Reverting this diff during peak hours may create a capacity SEV because we may not have the RE capacity to run every build action twice.
If this diff breaks a build, it means that build relies on some cached artifact from an old build that can no longer be reproduced, usually due to some pre-existing from non-hermeticity in the build. As a first step, you should try to fix the pre-existing issue rather than reverting this diff.

------

Add core build mode to the configuration name, so that these show up in "top level target platforms" column of buck2_builds, which we can then use to create a new "fbcode build modes" derived column for modefile-free

Reviewed By: JakobDegen

Differential Revision: D72881807

fbshipit-source-id: a55da3ec348780453defff4ae8834f989905cb64