http_{archive,file}: expose size_bytes + add a test info
Summary:
Like it says in the title. This way we can omit the http_head call entirely,
but we still validate that it's correct by exposing it as `TestInfo` with a local-only
test, which will force us to actually HTTP-materialize this.

This way, when you update a http_archive on CI, we actually validate it's
setup properly, but if you gave us `size_bytes`, and your thing it is in CAS, we
never access it again (unless you run buck2 test on your HTTP target).

This actually offers stronger validation than what we had before, which only verified
that your endpoint existed and could respond to a HEAD request (but we did it on every build),
whereas now we validate that the endpoint can return the right contents (but we only
check it on test).

A cleaner way to express this might be to make this some form of native test info that Buck2
can support "can I materialize this?",  but if you look at the testing below here it actually works
out quite neat as-is.

Notionally this allows download_file to basically behave like cas_artifact at runtime but fill the
CAS cache using the HTTP endpoint, while having test coverage to ensure that the artifacts don't
expire.

Context:

https://fb.workplace.com/groups/buck2users/posts/3927658267490525/?comment_id=3930312230558462

Reviewed By: cjhopman

Differential Revision: D72708673

fbshipit-source-id: 2bdb4978bfca9b0b40524dc99d3d600bed7b28ff