Implement tooling required for generating sources from .proto files
Summary:
D70602802 got reverted as it was not working with windows and caused RL tests to fail.
In this diff we are implementing slightly differennt ways to generate sources from .proto models in windows and linux/macOS.

credit for `gen_zip.py` goes to our one and only hick209

**Details**
We use `protoc` to generate Java source files from `.proto` models.
For models that use `gRPC`, `protoc` uses `protoc-gen-grpc-java` as a plugin.
The `protoc-gen-grpc-java` that we pass as a plugin should be an executable on the host that runs `protoc`.

- When we use `command_alias` for `protoc-gen-grpc-java`, it creates a wrapper around `.exe` and it doesn't work when we pass it to `protoc`. Instead, passing it as a `.exe` file works, and that is what I'm doing for Windows.
- The same approach is not suitable for macOS because it may use RE workers with Linux OS to run the `genrule`, and we have to use command_alias to properly use linux/macOS executables for `protoc-gen-grpc-java`.

I hope it helps clarify why we have slightly different ways of generating source from .proto files in Windows and Linux/macOS.

Reviewed By: IanChilds

Differential Revision: D72875297

fbshipit-source-id: 6bd6594b1d63400e19aad685b01a62e8b2bb43f0