Add a label for fdb to show warnings based on sanitizers
Summary:
We disable certain sanitizers under the debugger due to incompatibility. There is no easy way to detect exact sanitizers in build (like, is LSan enabled or only ASan).

So we aim to show warning if any sanitizer is enabled which is good enough. To show the warning we add `fdb` label.

Differential Revision: D72958981

fbshipit-source-id: 4f1a205d587cc715ccb6361cc6337198396ac226