Reland [buck2] prelude: Make fat platform handling in `command_alias` more explicit
Summary:
Original: D73008564

Diff wrt previous version:

```
 diff --git a/fbcode/buck2/prelude/command_alias.bzl b/fbcode/buck2/prelude/command_alias.bzl
 --- a/fbcode/buck2/prelude/command_alias.bzl
+++ b/fbcode/buck2/prelude/command_alias.bzl
@@ -12,7 +12,7 @@
     target_os = ctx.attrs._target_os_type[OsLookup]
     exec_is_windows = ctx.attrs._exec_os_type[OsLookup].os == Os("windows")

-    if target_os.os == Os("fat_mac_linux"):
+    if target_os.os == Os("fat_mac_linux") and len(ctx.attrs.platform_exe) > 0:
         variants = {
             "Darwin": _get_os_base(ctx, Os("macos")),
             "Linux": _get_os_base(ctx, Os("linux")),
```

The previous version of this diff caused a behavior change where we sometimes used trampolines in cases in which we previously did not. Not using trampolines seems to make it possible to hide a number of bugs in command_alias definitions:

This fixes

```
> buck2 build fbcode//mode/opt fbsource//xplat/third-party/node/node-zip:module -u fbcode//browserlab/proxy:asar
```

The actual bug is here:

https://www.internalfb.com/code/fbsource/[b2b3baa9a9b7]/xplat/third-party/node/build-deps/BUCK?lines=20

That's not a valid way to pass flags. The exact mechanism that makes this work today is not clear to me

Another thing that this fixes is caused by the bug here:

https://www.internalfb.com/code/fbsource/[b2b3baa9a9b7]/third-party/bison/3.8.2/BUCK?lines=25

This relies on missing shell escaping in `command_alias`. I am similarly not sure why it isn't broken today.

Reviewed By: dtolnay

Differential Revision: D73225444

fbshipit-source-id: 73d34973010cb3eef76e2f1272aed676b40acef7