command alias: Simplify output generation on Windows
Summary:
`command_alias` on Linux has to do this thing where it writes out relative paths like `"SOME_VAR/foo bar"`, and then runs sed to replace that with `"$SOME_VAR/foo bar"`, so that the quoting on the path does not affect the `$SOME_VAR`.

Windows used to do the same thing, however D45486970 observed that this is actually not necessary because we always do shell quoting (this is a bug of course, D73178171), and `%SOME_VAR%` isn't escaped by shell quoting.

However, it then went and removed this behavior for Windows exec platforms, instead of Windows target platforms, which is completely wrong, but it did sort of work because those will usually end up being the same.

This diff does the sane thing of removing this behavior for the windows target platforms regardless of exec platform choice. The previous behavior was one of the bugs that was not covered by existing tests in this stack.

Reviewed By: dtolnay

Differential Revision: D73292915

fbshipit-source-id: d8b5fcf58e1d3e7cb27545a6c96ff4b809ed59bd