Refactor UsedClasses plugin to output files line by line using URI
Summary:
### Context
In D70623505, I introduced ndjson files as a quick fix to address the incompatibility of the UsedClasses plugin with the option for the incremental compiler to run in multiple iterations within a single build. For more details, please refer to the diff description.

Upon further inspection of the kotlin-used-classes-tmp.txt content, I noticed that it essentially creates a map from a split URI. Given that merging text lines is significantly simpler than merging multiple JSON objects, I propose modifying the output of the UsedClasses plugin to adopt a one-line-per-URI format, consistent with the output format of the KSP/KAPT class usage tracker.

The map for later processing is generated in a subsequent step, alongside the output from the KSP/KAPT class usage trackers.

### This Diff
This diff refactors the UsedClasses plugin to output files line by line using URIs, instead of using ndjson format.

Differential Revision: D73110142

fbshipit-source-id: 153eb99de427db5ebf17b819d4ca1bb65a921371