Differentiate -Cmetadata by configured target (v2)
Summary:
NOTE: Part of this diff was previously landed in D73334841 and reverted in D73400841. This iteration normalizes away the "ovr_config//toolchain/python/constraints:python-site" constraint setting that accounted for the test failures in Monarch.

When bootstrapping rustc, various compiler crates depend on `cfg-if` in its default feature set, which depends on a sysroot including `std`, which depends on `cfg-if` in its "rustc-dep-of-std" feature set. The Buck target for `cfg-if` looks like this, with 2 reindeer universes for the library vs compiler stage of the bootstrap:

```lang=starlark
rust_bootstrap_library(
    name = "cfg-if-1.0.0",
    srcs = [":cfg-if-1.0.0.crate"],
    crate = "cfg_if",
    crate_root = "cfg-if-1.0.0.crate/src/lib.rs",
    edition = "2018",
    features = select({
        "//constraints:compiler": [],
        "//constraints:library": [
            "compiler_builtins",
            "core",
            "rustc-dep-of-std",
        ],
    }),
    named_deps = select({
        "//constraints:compiler": {},
        "//constraints:library": {
            "core": ":rustc-std-workspace-core-1.99.0",
        },
    }),
    visibility = [],
    deps = select({
        "//constraints:compiler": [],
        "//constraints:library": [":compiler_builtins-0.1.146"],
    }),
)
```

This fails to build because the compiler's build of `cfg-if` observes that a different `cfg-if` with the same `-Cmetadata` is already present in the sysroot.

```lang=text,counterexample
Action failed: rust//:cfg-if-1.0.0 (rustc metadata [pic])
error[E0519]: the current crate is indistinguishable from one of its dependencies: it has the same crate-name `cfg_if` and was compiled with the same `-C metadata` arguments, so this will result in symbol conflicts between the two

BUILD FAILED
Failed to build 'rust//:cfg-if-1.0.0 (rust//platforms/stage1:compiler#768747b6411b7118)'
```

Since the 2 crates are deliberately differently configured, we need to make them have different `-Cmetadata`.

Fbsource's `cfg-if` has not hit the same problem because it's vendored twice at different paths: the sysroot crates are built using **fbsource/xplat/rust/toolchain/sysroot/1.86.0/vendor/cfg-if-1.0.0** whereas the `cfg-if` used by everything else is built from **fbsource/third-party/rust/vendor/cfg-if-1.0.0**.

Reviewed By: JakobDegen

Differential Revision: D73414962

fbshipit-source-id: b2442ca15cfb7f9a21f8bba52d5a14e652d03b23